Make the display message code modular
	Until now everybody was adding type_* messages to the
message.flag structure in include/project.h. The problem with this
is that there are only 16bits in the message.flag field and thus
not a lot of apps can benefit from it at the same time.
	I fixed this by adding a type_msg and line_msg1 to the
message.flag struct and changing the relevant code in ezchronos.c
to:
1) Continue working with type_locked, etc.. which are used in
driver/*
2) Apps in logic/* should set message.flag.prepare=1 and
message.flag.type_msg = 1 and pick the correct line in
message.flag.line_msg1 to show a message in the display. By doing
this the display handler of the logic app will be called with a
DISPLAY_LINE_MESSAGE update.

NOTE: The logic/alarm was also updated to work with this patch.
I also tested the button lock, alarm, beep on/off, and wireless
functions in the watch, eerything works..

In summary, this patch allows a bunch of apps to benefit from the
message display update, without having to hack ezchronos.c and
and project.h, increasing modularity..