alarm: Updated with improved timer0 API
[rant on]
The alarm code is a fuc**ng mess! One can't even have hourly
beep and a alarm simultaneously!
And why was that the DST was inside a #ifdef ALARM?
I didn't get it!
[rant off]

If someone cares to cleanup the alarm code, one can just register
the timer0 when one of the hourly beeps or alarm is enabled and
unregister otherwise.. this will save CPU clk cycles.

Right now, the alarm code registers in the timer0 queue as soon
as the clock is initialized which is suboptimal and doesn't take
fully advantage of the new timer0 API.