alarm: Simplified code, cleanup, display improvements
* The timer is now only registered when needed, saving CPU cycles..
* Improved the way how hourly chimes and alarm are presented in the
screen. The alarm is now represented by the bell symbol while the
hourly chime is represented by the "))" symbol (2nd and 3nd beeper
LCD segments). This way, it is not only easier to know which types
of alarm are enabled but also quicker to cycle through them.
Because other "applications" might want to profit from the bell
symbol, I decided to turn it off while not in the alarm screen..
If people really want the bell shown during the clock screen,
let me know and I'll make the clock screen show both the "))"
and bell symbols, if the hourly chimes and alarm are respectively
enabled.
* I also simplified a lot a bunch of code, effectively reducing its
size. In the struct alarm, it was also possible to put pack most
of the fields into a byte which saves memory...

Well, test this patch! I believe you are going to like it!