rtca: Fix wrong RTC_A register access
The RTCMODE, RTCHOLD, .. macros are word size. I was using them with the
assumption they were byte size which was preventing the code to work
as expected.

The RTC is now enabled and working. If you want to test it, build and
flash the code into the chronos. The clock should be reset to 00:00:00
and count time from then on. (press up button for seeying the seconds).