rtca: Implement a time event interrupt handler
The RTC_A_ISR now calls a tevent_fn passing the type of event
as argument (hour, minute, ... changed). The logic/clock takes
advantage of this to only update certain parts of the display
when required. This was already being done before, the difference
is that the hour/minute change is now being provided by the RTC_A
module instead of being calculated by software.