rtca: Implement RTC_A based system_time
The system time variable is incremented using the RTCRDY interrupt
source, which happens every second. Since the RTCRDY interrupt was
enabled, I took advantage of it and cached the RTC registers values
in memory. By doing this, we no longer need to waste CPU cycles
while waiting for a register read ready, during a rtca_get_time().
This is because the cached values on RAM are guaranteed to be
correct.

Previously, the system_time was provided by logic/clock. This
old functionality was deleted and was replaced by the RTC systime.