Date implemented, small changes in RTC interrupt routine
Date is now using RTCA module. There are still small problems (like max day is always 31 regardless of month when adjusting date) Didn't have time to test it, will do later today.

I reordered code in rtca.c, eliminating the first if (there are only two reasons why we are in the interrupt routine: because of RTCRDY or RTCTEV) and refreshing only changed values in cache to speed up everything.

Signed-off-by: Aljaž Srebrnič <a2piratesoft@gmail.com>