rtca: Half migration of logic/clock to rtca
The hourly chime is now working with rtca.
Because the RTC_A hardware only allows keeping one alarm set at a time,
the alarm register will be used for alarm. The hourly chime is then using
the rtca_tevent_fn_t callback to receive hourly notifications on the
fly. Nonetheless, the use of rtca allowed considerable simplification
of the code.

The alarm is not working yet, it needs the RTC_A alarm implemented.
This will be done in the next commit.