display: KILL the display_update() callback
This patch kills the useless display_update() callback and all the crap
that goes with it. The display should always be refreshed in a regular
interval, with no exception. This simplifies a lot the code and removes
inter-cross-ultra-dependencies between modules.

Rational: If there is some code in logic/* that needs a higher
"refresh rate" then, we should increase the refresh rate for everyone
and not just a few modules. I propose a refresh rate of 100ms.
If everything is well implemented (like clock.c and date.c) unneeded
screen update requests ARE ignored by the module avoiding wasting
battery in useless LCD updates.

This is my second attempt at simplifying the display code.
Some modules like strength are not working. They will be fixed when
the display code is more modular and simple.