New display system: First commit.
This commit comes as consequence of the decisions taken in the irc
channel.

In this commit I removed all traces of message.* and display.* global
flags. There is a lot of stuff that stopped working because of this,
including button lock, buzzer mute, etc..

The good new is that I actually removed a bunch of code, this is going
to simplify development from now on and will also make firmware size
smaller.

I also removed the double-menu functionality and killed the global
display_update function. Display updates are now fully done by
modules with no interference whatsoever from the menu system.
A new set of callback functions were added to the menu_add_item
function.