clock: Implement date.
Date is now displayed in clock module. The date is displayed in the
second row, so the total output of clock is:

00:00
01-01

where the first line is HH:MM and the second line is DD:MM.
The clock module was removed from the tree since it is not needed
anymore.

Setting the date also works the same way time is set. There is one
problem though, to calculate the max nr of days in a month we need
the year! For this reason, we have to find a way to show current
year and let the user change it.