drivers/display: Support state bitfield in write_lcd_mem()
The family of functions display_char{s}() and display_symbol() all
use write_lcd_mem() to write segment data and blinking data to memory.
The choice to enable or disable a segment and to enable or disable
blinking of that segment is provided by the status argument. However,
because the status argument was NOT a bitfield, the programmer could not
turn on only one segment and leave its blinking status untouched. This
was a huge limitation which was pain in the ass.

This patch makes the write_lcd_mem function support a bitfield status
and allow the programmer to touch only the LCD memory, only the
blinking memory or BOTH.

NOTE: All modules were updated to use the new functionality.

NOTE: This patch also fixes a bug in display_chars() which incorrectly assumed
that the str argument could be NULL. Well it could NOT and I wonder why
the chronos didn't crash so far..
You can now set str=NULL, display_chars() will automatically assume
str is a vector of '8' characters.