Make cblist functions universal
Having a queue of callbacks and a set of functions that add and
remove callbacks from that queue is needed by several drivers.
For that reason, I moved the cblist_register/unregister functions
to ezchronos.c so that all the drivers can reuse the code.

I also updated rtca and timer drivers to use the new functions. This
should save us some memory..

As a side note, I also implemented the 10Hz timer. The 1Hz and 10Hz
timer queues are exported in timer.h, relevant modules wanting to
listen for those timer events should use the cblist_register/unregister
functions with a timer queue.