Implement a system message bus
Original ezchronos worked by having the execution inside the mainloop
pause at LPM3. The watch would then remain in sleep until a interrupt
routine happens. The interrupt routine would then do some work
and set a global flag telling that something new is available. At
exit the interrupt routine would then call _BIC_SR_IRQ(LPM3_bits)
which tells the MCU to exit from LPM3. At this time, the execution
is resumed at the mainloop where it checks for the global flags and
do the remaining work.. until it goes into LPM3 again.

I only figured this out by now.. (I know I should have checked how
low power modes work here first). But in fact, nothing was lost. The
original plan was to have the interrupt routines do all the work
directly through a system of callbacks. It works, but the ezchronos way
has several advantages over this:

1) Because mostly all the work is done in the mainloop and the watchdog
is serviced in the mainloop. If execution gets stuck somewhere there,
then the watchdog can do what it does best. Otherwise, in the case
where interrupts do most of the work, if execution gets stuck in a
interrupt, then the watchdog wouldn't know (unless its the interrupt
routine that wakes up the system).

2) Interrupts shouldn't do a lot of processing anyway.

3) Acording to the datasheet, exiting from LPM3 takes some time.
Consolidating several independent wakeups from all the interrupts
into a single longer wakeup improves efficiency.

By this reason, I migrated all the code into a new system message bus.
The concept is very simple and similar to other message bus (also
somewhat similar to original ezchronos, but more elegant and cleaner).
Each interrupt does minimal work and updates a status variable exported
by the driver where the interrupt belongs to. The mainloop then checks
that status variable and all the others belonging to other drivers and
builds a message. A message is just a bitfield with all the events that
happened very soon, for example, RTC minute and a 10Hz event. That
message is broadcasted to all the listerers in the bus.
A listener is everyone that calls sys_messagebus_register(). When
registering in the bus, the listener can select to be notified only
for a few types of messages.

Finally, I just want to leave a remark. The way the system message bus
is implemented, it will only broadcast the latest events. If, say a
100Hz timer interrupt happens more than one time while the message bus is
broadcasting a message, then the all 100Hz events except the last one
are lost. This is not a bug, it is a feature. If events are lost, then
its because the watch was not fast enough to handle them anyway.