Super big mega cleanup
* SimpliciTI is gone. The sources are still around but all references
to it were removed. Basically, syncing time with your computer is not
working anymore. Wireless update is a different matter and independent
from simpliciTI so it still works.

* Removed all references to logic/*. Stuff in logic is not up to date
with current system, so they need to be ported. They will be moved
from there to modules/ as they are ported.

* Removed a lot of other stuff that superfulous and take a lot of RAM.
If in the future you notice something is missing, then you should look
at this commit first :)

The good news is that the firmware size is getting really really small
and the entire system getting very clean. We should start fixing broken
stuff soon.