drivers/ports: Use falling edge interrupts to detect button releases
This patch reduces in about 50% the code size of the buttons driver.
It also doesn't need to pool the buttons for long button presses anymore.
I also feel the code responds better to user input than before.

This is how it works, each input PORT can be configured to raise
interrupts on rising or falling edge. When you press a button,
it raises a rising edge interrupt and the driver saves the time
when you pressed it. The driver then switches the next interrupt for
that button to falling edge so we get an interrupt when you release
the button. When you actually release the button, a interrupt is raised
and it checks how long you toke to release it. That's how it
distinguishes between a long and short button press. The driver then
switches the button to rising edge interrupts again.

I found that was not necessary to debounce the input. I actually like
the fast response that it gives by not spending time debouncing.
Also, because the system is actually waiting for you to release the
button, you don't get any feedback that you pressed long enough. By
I think this is not an issue, I got used to it in a few minutes.