drivers/ports: Detect simultaneous button presses
Detecting long button presses has a problem, it only works if the 10hz
timer is active. If for some reason (debugging) the timer is disabled,
then we cannot do the long num press to launch rfbsl, forcing us to
disassemble the watch and use the programmer.

Detecting simultaneous button presses is free, no timer is needed. For
that reason I changed the way rfbsl is launched from long num press
to simultaneous up&down press. This way RFBSL always works, even
when no timers are active in the system.