Implement wireless updating at boot
This patch implements wireless updating during chronos initialization.
After a reset and before entering the main() function, the watch checks
if the up&down buttons are simultaneously pressed. If they are, the
wireless updater is launched.

Note: There are no delays during up&down button press detection which
means that *you already have to be pressing the buttons before the
watch resets*

Why is this useful?
As Jeffrey Schwartz suggested in the mailing list, this can be useful if
the watch is reseting in a predictable fashion (like a crash followed by
a watchdog reset). Just start pressing the relevant buttons before
the watch resets and that should avoid you from recurring to the
screwdriver.

How to test this?
In main() in openchronos.c, just comment the following line:
WDTCTL = (WDTCTL & 0xff) | WDTPW | WDTCNTCL;

That will prevent the watchdog from being serviced which makes it
reset the clock every 16 seconds.