Rough Nokia Buzzer support
Completely rewritten buzzer.[ch]. The central function is now
buzzer_play, that accept an array of notes defined thusly:
- first 4 MSB are the pitch
- next 2 bits specifies octave
- last 10 bits define duration
There are also 2 "meta" notes:
0x___0 defines a rest (silence)
0x___F is a "stop" note, ending a melody sequence.

TI license was removed as the file was completely rewritten.

Keep in mind that the current implementation does not power save; the
watch is active and "locked" for the whole duration of a melody. This is
of course unacceptable, and will be solved soon by using the timer
module.