Add button debouncing to the button interrupt handler
This patch makes the watch ignore all button presses that were quicker
than a specified time interval. Because this uses the 10Hz timer to
measure the time interval, a fail-safe mechanism was implemented
to allow the buttons to still work even when the timer is disabled.
In this case, the buttons won't be debounced and only short button
presses will be detected.