Increase Timer0 CCR0 frequency from 10Hz to 20Hz
This allows the "deboucing" delay implemented in
5a25c773c777db81baf7ffc86d7980e66df9e73b to detect shorter button
presses, improving input usability.

Since at the moment only the ports driver was using the 10Hz timer,
there is no harm in increasing its frequency to 20Hz, with the exception
of a somewhat negligible battery usage increase. Future modules might
also benefit from this frequency, specially for measurements of small
periods of time.