Implement item based menu_editmode
Looking at both alarm and clock modules, I was still not happy
with how the code had to be written to properly use menu_editmode.
In special, the looping through all edit items had to be done manually
by all modules. Moreover, all modules also needed to conditionally
chose the item to blink/set. All of this simple tasks are always repeated
between modules implementing menu_editmode. For this reason, I merged
them into the openchronos menu_editmode functionality so that the tasks
are performed automatically by the openchronos core.
For now, this functionality is taking a hundred bytes more than without
it, which I think is fine since it really improves code readability.

The clock and alarm modules were updated to use the new API.
The annotated doxygen documentation was also updated.