Improve virtual screens functionality: active screen
This patch adds support for active screens on the virtual screens
framework. Until now, contents of each virtual screen needed to
be manually copied to the real screen to ensure the real screen was up to
date. This patch however, allows setting one of the virtual screens
as "active". All display updates to the active virtual screen will
imediatly be visible in the real screen.

Here is how it works:
A developer creates N virtual screens using lcd_screens_create().
From the N screens, screen 0 will be automatically selected as
active. All display updates to that screen will be directly visible
in the real screen. Modules can also display data on the non-active
virtual screens, however that data will not be visible on the scren
but it will be saved. At a certain point (for example when the user
presses the NUM button) the lcd_screen_activate() function can be
called to switch the active screen. When a new screen is made
active, the contents of the real screen are automatically updated.

See the alarm or clock module for an example.

This patch reduces the firmware size in about hundred bytes and
reduces by one the total number of virtual screens in use at all times.