Added python script to create ringtones
The support is very crude, and it should probably be merged in the
Chronos tool, but here it is.

Usage: rtttl2bin [string]

Outputs the binary ringtone directly in C format, so it can be
copy-pasted directly into OpenChronos-ng source code. It dies with an
error if a note is too long to be stored, I must still think about what
to do in this case (the longest note representable is 1023ms long).