Menu mode improvement: b/fwd module selection and name display
This patch allows the user to navigate between modules using the
up/down buttons. Also, a MENU string is displayed on first row
and the selected module name is displayed in the second row.

Hopefully this menu will make it easier to navigate in the future,
when lots of modules are compiled in.