Implement pseudo sprintf function based on _itoa and friends
char *_sprintf(const char *fmt, int16_t n);

This function returns a pointer to the string containing the number
n formatted according to fmt. The format is NOT cmpatible with stdio's format.
Examples:
// returns " 8"
_sprintf("%2u", 8);

// returns "0020"
_sprintf("%04u", 20);

// returns "-048"
_sprintf("%03s", -48);

// returns " 048"
_sprintf("%03s", 48);

// returns "0xff"
_sprintf("0x%02", 0xff);

// returns "st1x"
_sprintf("st%1ux", 1)

WARNING:
You must always specify the number of digits or bad results
will happen! "%u" formats are not allowed!

This patch reduces in about hundred bytes the size of the firmware.