Implement temporal mean filtering on temperature driver
This commit makes the driver average the last 4 measurements, to
smooth out quick oscillations.
The temperature driver was also rearranged to provide direct
readings on either degrees C or F.
The offset parameter was moved to influence the raw sensor value
instead of the temperature (C/F), making it independent of what
displaying units were selected.

I was expecting the code to decrease but it was not the case. We
might consider trying to optimize this further in the future. Maybe
by only storing the first 8 LSB in the adcresult vector?