Implement boot menu
Starting from now, there is no RFBSL module anymore. There is
however a "reset" module that you can use to .. reset the watch.

When the watch boots, it remains in LPM3 until a button is pressed.
If the button is BACKLIGHT, then it will launch RFBSL and allows
you to flash the watch over the air. If any other button is pressed,
it continues execution into main openchronos.

Just a note, I moved some uC initialization routines to boot.c.