Rework of buttons driver and input layer
I used the watch for a day and tried to figure out what I could
do to make it more user friendly. The first thing that came to
my mind was that I always use the thumb + indicator fingers
to hold the watch, while one of them is pressing a button.
This means that when the thumb is doing a long STAR button press,
the indicator finger is holding a non-button zone of the watch
to compensate for the force of the thumb. This gave me the idea
that maybe instead of a long STAR press, a simultaneous press of
STAR + UP would be more handy. The same applies for long NUM
which was replaced by NUM + DOWN.
This also has the side benefit of simplifying the code a lot.

In a slightly related notice, I finally implemented the long
waited button auto-repeat (or continuous press as you want
to call it). Basically, if you are in edit mode (say
setting the time or alarm), you can keep pressing UP/DOWN to
increment the values.

TODO: Fix bug where number is incremented after entering
edit-mode (thanks Petre Rodan for reporting).